0719 yuluo/add unit test

+++ manager/src/main/java/org/apache/hertzbeat/manager/service/impl/TagServiceImpl.java
@@ -24,9 +24,12 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.hertzbeat.common.entity.manager.Monitor;
 import org.apache.hertzbeat.common.entity.manager.Tag;
+import org.apache.hertzbeat.common.support.exception.CommonException;
 import org.apache.hertzbeat.manager.dao.TagDao;
+import org.apache.hertzbeat.manager.dao.TagMonitorBindDao;
 import org.apache.hertzbeat.manager.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -36,6 +39,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
+ * Tag service implementation.
  */
 @Service
 @Transactional(rollbackFor = Exception.class)
@@ -44,6 +48,9 @@ public class TagServiceImpl implements TagService {
 
     @Autowired
     private TagDao tagDao;
+
+    @Autowired
+    private TagMonitorBindDao tagMonitorBindDao;
 
     @Override
     public void addTags(List<Tag> tags) {
@@ -54,6 +61,7 @@ public class TagServiceImpl implements TagService {
     public void modifyTag(Tag tag) {
         Optional<Tag> tagOptional = tagDao.findById(tag.getId());
         if (tagOptional.isPresent()) {
+            tag.setTagValue(StringUtils.isEmpty(tag.getTagValue()) ? null : tag.getTagValue());
             tagDao.save(tag);
         } else {
             throw new IllegalArgumentException("The tag is not existed");
@@ -67,6 +75,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void deleteTags(HashSet<Long> ids) {
+        if (tagMonitorBindDao.countByTagIdIn(ids) != 0) {
+            throw new CommonException("The tag is in use and cannot be deleted.");
+        }
         tagDao.deleteTagsByIdIn(ids);
     }
